audio_synchronization/concurrent_slab: reimplement on top of other primitives
It should now be obviously safe (indeed it contains no unsafe itself) and
hopefully obviously correct (since we couldn't model the old one with loom, but the new one
is at least built on loom-testable primitives or crates we didn't write like arc_swap).

+++ crates/synthizer/src/data_structures/object_pool.rs
@@ -7,7 +7,7 @@ use ahash::{HashMap, HashMapExt};
 use audio_synchronization::concurrent_slab::*;
 
 /// Initial capacity for each slab in the pool.
-const INITIAL_CAPACITY: u32 = 100;
+const INITIAL_CAPACITY: usize = 100;
 
 /// Manages a collection of [SlabHandle]s, each for a different kind of object.
 pub(crate) struct ObjectPool {
